Akhirnya selesai jg rapihin Eunomia sama Zeus

+++ database/migrations/002_create_zeus_table.php
@@ -14,28 +14,36 @@ return new class extends Migration
      */
     public function up(): void
     {
+        // Create the database if it doesn't exist
         DB::connection('mysql')->statement('CREATE DATABASE IF NOT EXISTS '.$this->connection);
 
-        Schema::create('profiles', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('users_id')->constrained(table: 'eunomia.users')->onDelete('cascade');
-            $table->string('name');
-            $table->string('place_of_birth')->nullable();
-            $table->date('date_of_birth');
-            $table->tinyInteger('gender')->default(0);
-            $table->tinyInteger('blood_type')->default(0);
-            $table->tinyInteger('identity_type')->default(0);
-            $table->string('identity_number', 16)->nullable();
-            $table->timestamps();
-        });
-        Schema::create('notifications', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('profiles_id')->constrained('zeus.profiles')->onDelete('cascade');
-            $table->text('title')->nullable();
-            $table->text('body');
-            $table->boolean('opened')->default(0);
-            $table->timestamps();
-        });
+        // Check and create 'profiles' table
+        if (!Schema::hasTable('profiles')) {
+            Schema::create('profiles', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('users_id')->constrained(table: 'eunomia.users')->onDelete('cascade');
+                $table->string('name');
+                $table->string('place_of_birth')->nullable();
+                $table->date('date_of_birth');
+                $table->tinyInteger('gender')->default(0);
+                $table->tinyInteger('blood_type')->default(0);
+                $table->tinyInteger('identity_type')->default(0);
+                $table->string('identity_number', 16)->nullable();
+                $table->timestamps();
+            });
+        }
+
+        // Check and create 'notifications' table
+        if (!Schema::hasTable('notifications')) {
+            Schema::create('notifications', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('profiles_id')->constrained('zeus.profiles')->onDelete('cascade');
+                $table->text('title')->nullable();
+                $table->text('body');
+                $table->boolean('opened')->default(0);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
